Fb Login

+++ vagrant/catalog/finalproject.py
@@ -3,7 +3,7 @@ from flask_bootstrap import __version__ as FLASK_BOOTSTRAP_VERSION
 from flask_bootstrap import Bootstrap
 from flask_nav.elements import Navbar, View, Subgroup, Link, Text, Separator
 from markupsafe import escape
-from database_setup import Base, Restaurant, MenuItem
+from database_setup import Base, Restaurant, MenuItem, User
 from sqlalchemy import create_engine
 from sqlalchemy.orm import sessionmaker
 from forms import RestaurantForm, RestaurantMenuForm
@@ -36,6 +36,73 @@ def showLogin():
       string.digits)for x in xrange(32))
     login_session['state'] = state
     return render_template('login.html', STATE=state)
+
+
+@app.route('/fbconnect', methods=['POST'])
+def fbconnect():
+    if request.args.get('state') != login_session['state']:
+        response = make_response(json.dumps('Invalid state parameter.'), 401)
+        response.headers['Content-Type'] = 'application/json'
+        return response
+    access_token = request.data
+    print "access token received %s " % access_token
+
+    app_id = json.loads(open('fb_client_secrets.json', 'r').read())[
+        'web']['app_id']
+    app_secret = json.loads(
+        open('fb_client_secrets.json', 'r').read())['web']['app_secret']
+    url = 'https://graph.facebook.com/oauth/access_token?grant_type=fb_exchange_token&client_id=%s&client_secret=%s&fb_exchange_token=%s' % (
+        app_id, app_secret, access_token)
+    h = httplib2.Http()
+    result = h.request(url, 'GET')[1]
+
+    # Use token to get user info from API
+    userinfo_url = "https://graph.facebook.com/v2.4/me"
+    # strip expire tag from access token
+    token = result.split("&")[0]
+
+
+    url = 'https://graph.facebook.com/v2.4/me?%s&fields=name,id,email' % token
+    h = httplib2.Http()
+    result = h.request(url, 'GET')[1]
+    # print "url sent for API access:%s"% url
+    # print "API JSON result: %s" % result
+    data = json.loads(result)
+    login_session['provider'] = 'facebook'
+    login_session['username'] = data["name"]
+    login_session['email'] = data["email"]
+    login_session['facebook_id'] = data["id"]
+
+    # The token must be stored in the login_session in order to properly logout, let's strip out the information before the equals sign in our token
+    stored_token = token.split("=")[1]
+    login_session['access_token'] = stored_token
+
+    # Get user picture
+    url = 'https://graph.facebook.com/v2.4/me/picture?%s&redirect=0&height=200&width=200' % token
+    h = httplib2.Http()
+    result = h.request(url, 'GET')[1]
+    data = json.loads(result)
+
+    login_session['picture'] = data["data"]["url"]
+
+    # see if user exists
+    user_id = get_user_id(login_session['email'])
+    if not user_id:
+        user_id = create_user(login_session)
+    login_session['user_id'] = user_id
+
+    output = ''
+    output += '<h1>Welcome, '
+    output += login_session['username']
+
+    output += '!</h1>'
+    output += '<img src="'
+    output += login_session['picture']
+    output += ' " style = "width: 300px; height: 300px;border-radius: 150px;-webkit-border-radius: 150px;-moz-border-radius: 150px;"> '
+
+    flash("Now logged in as %s" % login_session['username'])
+    return output
+
 
 @app.route('/gconnect', methods=['POST'])
 def gconnect():
@@ -155,20 +222,27 @@ def gdisconnect():
 @app.route('/restaurants')
 def restaurant_index():
     restaurants = session.query(Restaurant).all()
-    return render_template('restaurant.html',restaurants=restaurants)
+    if 'username' not in login_session:
+        return render_template('publicrestaurant.html',restaurants=restaurants)
+    else:
+        return render_template('restaurant.html',restaurants=restaurants)
 
 @app.route('/restaurant/<int:restaurant_id>/')
 def restaurant_menu(restaurant_id):
     restaurant = session.query(Restaurant).filter_by(id=restaurant_id).one()
+    creator = get_user_info(restaurant.user_id)
     items = session.query(MenuItem).filter_by(restaurant_id=restaurant.id)
-    return render_template('menu.html', restaurant = restaurant, items = items)
+    if 'username' in login_session and creator.id == login_session['user_id']:
+        return render_template('menu.html', restaurant = restaurant, items = items)
+    else:
+        return render_template('publicmenu.html', restaurant = restaurant, items = items)
 
 @app.route('/restaurant/new', methods=['GET','POST'])
 def new_restaurant():
     form = RestaurantForm()
     if request.method == 'POST':
-        name= request.form['name']
-        restaurant = Restaurant(name = name)
+        name = request.form['name']
+        restaurant = Restaurant(name=name, user_id=login_session['user_id'])
         session.add(restaurant)
         session.commit()
         flash('Restaurant has been created successfully!','success')
@@ -179,8 +253,13 @@ def new_restaurant():
 @app.route('/restaurant/<int:restaurant_id>/edit', methods=['GET','POST'])
 def edit_restaurant(restaurant_id):
     restaurant = session.query(Restaurant).filter_by(id=restaurant_id).one()
+    creator = get_user_info(restaurant.user_id)
     form = RestaurantForm()
     form.name.data = restaurant.name
+    if 'username' not in login_session:
+        return redirect('/login')
+    if restaurant.user_id != login_session['user_id']:
+        return "<script>function myFunction() {alert('You are not authorized to edit this restaurant. Please create your own restaurant in order to edit.'); window.location.href = '/' }</script><body onload='myFunction()''>"
     if request.method == 'POST':
         name = request.form['name']
         restaurant.name = name
@@ -194,6 +273,11 @@ def edit_restaurant(restaurant_id):
 @app.route('/restaurant/<int:restaurant_id>/delete', methods=['GET','POST'])
 def delete_restaurant(restaurant_id):
     restaurant = session.query(Restaurant).filter_by(id=restaurant_id).one()
+    creator = get_user_info(restaurant.user_id)
+    if 'username' not in login_session:
+        return redirect('/login')
+    if restaurant.user_id != login_session['user_id']:
+        return "<script>function myFunction() {alert('You are not authorized to delete this restaurant. Please create your own restaurant in order to delete.'); window.location.href = '/' }</script><body onload='myFunction()''>"
     if request.method == 'POST':
         session.delete(restaurant)
         session.commit()
@@ -208,6 +292,13 @@ def new_menu_item(restaurant_id):
     description = None
     price = None
     form = RestaurantMenuForm()
+    if 'username' not in login_session:
+        return redirect('/login')
+    restaurant = session.query(Restaurant).filter_by(id=restaurant_id).one()
+    if login_session['user_id'] != restaurant.user_id:
+        script = "<script>function myFunction() {alert('You are not authorized to add menu items to this restaurant. Please create your own restaurant in order to add items.'); window.location.href='/restaurant/%s';}</script><body onload='myFunction()''>" % (str(restaurant_id))
+        print script
+        return script
     if form.validate_on_submit():
         name = form.name.data
         description = form.description.data
@@ -221,7 +312,12 @@ def new_menu_item(restaurant_id):
 
 @app.route('/restaurant/<int:restaurant_id>/menuitem/<int:menu_id>/edit', methods=['GET','POST'])
 def edit_menu_item(restaurant_id, menu_id):
+    if 'username' not in login_session:
+        return redirect('/login')
+    restaurant = session.query(Restaurant).filter_by(id=restaurant_id).one()
     item = session.query(MenuItem).filter_by(id=menu_id, restaurant_id=restaurant_id).one()
+    if login_session['user_id'] != restaurant.user_id:
+        return "<script>function myFunction() {alert('You are not authorized to edit menu items to this restaurant. Please create your own restaurant in order to edit items.'); window.location.href='/restaurant/%s';}</script><body onload='myFunction()''>" % (restaurant_id)
     form = RestaurantMenuForm(obj=item)
     if form.validate_on_submit():
         name = form.name.data
@@ -238,7 +334,12 @@ def edit_menu_item(restaurant_id, menu_id):
 
 @app.route('/restaurant/<int:restaurant_id>/menuitem/<int:menu_id>/delete', methods=['GET','POST'])
 def delete_menu_item(restaurant_id, menu_id):
+    if 'username' not in login_session:
+        return redirect('/login')
+    restaurant = session.query(Restaurant).filter_by(id=restaurant_id).one()
     item = session.query(MenuItem).filter_by(id=menu_id, restaurant_id=restaurant_id).one()
+    if login_session['user_id'] != restaurant.user_id:
+        return "<script>function myFunction() {alert('You are not authorized to edit menu items to this restaurant. Please create your own restaurant in order to edit items.'); window.location.href='/restaurant/%s';}</script><body onload='myFunction()''>" % (restaurant_id)
     if request.method == 'POST':
         session.delete(item)
         session.commit()
@@ -248,7 +349,7 @@ def delete_menu_item(restaurant_id, menu_id):
         return render_template('delete_menu_item.html', restaurant_id = restaurant_id, item = item)
 
 def create_user(login_session):
-    newUser = User(name=login_session['name'], email=login_session['email'],
+    newUser = User(name=login_session['username'], email=login_session['email'],
                 picture=login_session['picture'])
     session.add(newUser)
     session.commit()
